requirements.txt
# Core RAG
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
faiss-cpu>=1.12.0
sentence-transformers==3.0.1

# Hallucination Detection
transformers==4.44.2
torch==2.4.0
scipy==1.14.1
numpy==1.26.4

# Evaluation
ragas==0.1.21
datasets==2.21.0

# Data Ingestion
arxiv==2.1.3
pypdf==4.3.1

# Dashboard
streamlit==1.38.0
plotly==5.24.1
pandas==2.2.2

# Utilities
python-dotenv==1.0.1
pydantic==2.9.1
tenacity==8.5.0
tqdm==4.66.5
